Renamed module file to state-utils and other fixes
