+++ lib/cluster/services/cluster/backends/native.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const Queue = require('@terascope/queue');
 const parseError = require('@terascope/error-parser');
-const stateUtils = require('./stateUtils');
+const stateUtils = require('./state-utils');
 
 /*
  Execution Life Cycle for _status
